+++ src/types.jl
@@ -99,6 +99,7 @@ mutable struct NumberConstraintTypes
     inequality::Int
     notequal::Int
     alldifferent::Int
+    element::Int
     table::Int
     indicator::Int
     reified::Int
@@ -130,6 +131,17 @@ Base.copy(A::AllDifferentSetInternal) = AllDifferentSetInternal(A.dimension)
 
 struct AllDifferentSet <: JuMP.AbstractVectorSet end
 JuMP.moi_set(::AllDifferentSet, dim) = AllDifferentSetInternal(dim)
+
+struct Element1DConstInner <: MOI.AbstractVectorSet
+    dimension::Int
+    array::Vector{Int}
+end
+Base.copy(E::Element1DConstInner) = Element1DConstInner(E.dimension, E.array)
+
+struct Element1DConst <: JuMP.AbstractVectorSet 
+    array::Vector{Int}
+end
+JuMP.moi_set(E::Element1DConst, dim) = Element1DConstInner(dim, E.array)
 
 struct TableSetInternal <: MOI.AbstractVectorSet
     dimension::Int
@@ -322,6 +334,7 @@ struct ReifiedSet{A,F,S<:Union{MOI.AbstractScalarSet,MOI.AbstractVectorSet}} <:
     set::S
     dimension::Int
 end
+ReifiedSet{A,F}(set::S) where {A,F,S} = ReifiedSet{A,F,S}(set, 1+MOI.dimension(set))
 Base.copy(R::ReifiedSet{A,F,S}) where {A,F,S} = ReifiedSet{A,F,S}(R.set, R.dimension)
 
 abstract type AbstractBoolSet{
@@ -437,6 +450,15 @@ mutable struct AllDifferentConstraint <: Constraint
     scc_init::SCCInit
     # corresponds to `in_all_different`: Saves the constraint idxs where all variables are part of this alldifferent constraint
     sub_constraint_idxs::Vector{Int}
+end
+
+mutable struct Element1DConstConstraint <: Constraint
+    std::ConstraintInternals
+    zSupp::Vector{Int} # number of possibilities for setting z (in z == T[y])
+    z::Variable
+    y::Variable
+    z_changes_ptr::Int # the start pointer for checking z.changes
+    y_changes_ptr::Int # the start pointer for checking z.changes
 end
 
 abstract type BoolConstraint{C1<:Constraint,C2<:Constraint} <: Constraint end